// wink-plugin-peripherals/builtin/mono_oled/1.0.0/src/variants.generated.ts
// [AUTO-GENERATED by @wink-ai/unisim-ui/generator. DO NOT EDIT MANUALLY.]
// Source: wokwi-ssd1306
import type { PinsOverlayMap } from '@wink-ai/unisim-ui';

export const MONO_OLED_AUTOGEN_SIZE = Object.freeze({
  width: 150,
  height: 116,
});

export const MONO_OLED_AUTOGEN_OVERLAY: Readonly<PinsOverlayMap> = Object.freeze({
  DATA: Object.freeze({
    relX: 37,
    relY: 13,
    wireNet: 'primary' as const,
    defaultConnection: null,
    required: true,
  }),
  CLK: Object.freeze({
    relX: 46,
    relY: 13,
    wireNet: 'secondary' as const,
    defaultConnection: null,
    required: false,
  }),
  DC: Object.freeze({
    relX: 55,
    relY: 13,
    wireNet: 'secondary' as const,
    defaultConnection: null,
    required: false,
  }),
  RST: Object.freeze({
    relX: 65,
    relY: 13,
    wireNet: 'secondary' as const,
    defaultConnection: null,
    required: false,
  }),
  CS: Object.freeze({
    relX: 75,
    relY: 13,
    wireNet: 'secondary' as const,
    defaultConnection: null,
    required: false,
  }),
  '3V3': Object.freeze({
    relX: 84,
    relY: 13,
    wireNet: 'vcc' as const,
    defaultConnection: '3V3' as const,
    required: false,
  }),
  VIN: Object.freeze({
    relX: 94,
    relY: 13,
    wireNet: 'vcc' as const,
    defaultConnection: 'VCC' as const,
    required: false,
  }),
  GND: Object.freeze({
    relX: 104,
    relY: 12,
    wireNet: 'gnd' as const,
    defaultConnection: 'GND' as const,
    required: false,
  }),
});